Move and refactor coord path positions
